用户管理完成

+++ src/main/java/com/goufn/permission/mapper/MenuMapper.java
@@ -18,7 +18,7 @@ public interface MenuMapper extends BaseMapper<SysMenu> {
      * @param userId 状态
      * @return set
      */
-    Set<String> findPermsByUserId(Integer userId);
+    Set<String> findPermsByUserId(Long userId);
 
     /**
      * 查询用户按钮
